requirements.txt
configparser
multiping
psutil
python-pushover
wakeonlan
pymysql